Update the addon to use the ActionButton API so that the button shows up in newer Firefox versions
The addon button doesn't show up in newer releases of Firefox at all. I have tried on Windows and Linux and I cannot still see the addon button to use it. See issue #11. I have updated the addon code to use the ActionButton API so that it works on newer Firefox versions.